Make tests in t-tap pass once again
The tests link to common/shared.c, which they're more or less meant to
do, but not to common/downtime.c, which they're not meant to. However,
putting global variables in common/downtime.c that are also accessed
elsewhere means we can't build the code without also linking in
common/downtime.c. That's a bit of a hassle, so let's move the
next_downtime_id variable to common/shared.c where the rest of them
live (well, with some few exceptions, but no rule's perfect if
everyone follows them).

Signed-off-by: Andreas Ericsson <ae@op5.se>