commands: disregard newlines in commands
The command delimiter should be decided by the source of the command (be
it a FIFO, a regular file or a custom worker.)

In other words, the input to process_external_command1 should be assumed
to be already delimited so that the end of the string denotes the end of
the command.

There is, to my knowledge, no use case for being able to pass several
commands to the function so there is no need to create this artificial
restriction on, for instance, the output of a passive check result, or
a host or service comment.

Signed-off-by: Anton Lofgren <alofgren@op5.com>