downtime: Fix dangling downtime pointers
Given a  retention data file (at least certain retention files,
probably relating to the order of the downtimes therein - I believe
they became inversed at some point), with a bunch of downtimes that
are all already expired, valgrind would report memory errors on startup,
and shortly thereafter we would get stuck in an infinite loop while
looking for triggered downtimes.

It turns out the prev pointers are sloppily kept. While the new
downtime insertion and removal code assumed that prev pointer would
point to the previous element in the list, in practice it would just
point at something that probably at some point was a downtime object
but at this time might either be free memory or the other end of the
list. When downtimes are later deleted, they will try to patch up the
linked list by using said prev pointer, leading to the amusing (or not,
if you actually want your monitoring system to monitor) situation
described above.

Now next and prev appears to be set to the next and the previous
element - or null, if we're at the end of the list - throughout the
list, which takes care of the issue.

Reported-by: Aron Lander <alander@op5.com>
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>