base/utils/janitor: avoid overflow compiler warning
in order to be able to compile with -Wall -Werror, we need to make sure
that we don't get an integer overflow when adding to a signed type,
since this might lead to undefined behaviour (unless we compile with
-fwrapv which would guarantee an overflow to result in a wraparound
which would also be ok(-ish)). I also changed the magic constant 84800
(almost the number of seconds in a day) to 86400 (number of seconds in
a day), so that it works as intended, as well as defining it to avoid
future confusion.

Signed-off-by: Anton Lofgren <alofgren@op5.com>