checks: don't require neb modules to redo service alert calculations
It's pretty unlikely that they want to do something different (although
that's still possible.) Nebmods that want to report/record host and
service alerts can now do so, rather effortlessly.

This patch adds NEBATTR_CHECK_ALERT to signify alert status of a check.

This allows us to avoid duplicating alert controls for modules which
just want to record alert information on their own.

Unfortunately, this requires some ABI breakage (in checks:handle_host_status)
which for some unknown reason is not declared static, despite not being
used anywhere outside the compilation unit, in order to gain access to the
alert_recorded information. It would be _possible_ to work around this
by recording the information to a global instead, but I really feel we
have enough of those.

Change-Id: I2667d093611eb1e8a91953d4713cc68c6138076d
Signed-off-by: Anton Lofgren <alofgren@op5.com>
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>