events: Never run checks ahead of schedule
This isn't a realtime system, where the scheduler is supposed to run
events before the deadline. In fact, the exact opposite is true - events
are expected to run as soon as possible, *after* the deadline.

By removing the opportunistic ahead-of-schedule execution, we get rid of
a nasty bug where the log rotation would happen until the log to rotate
was gone completely. The bug worked as such:
 * The log rotation is supposed to happen, at midnight, November 1st.
 * 100ms before midnight, we have spare time, and rotates the log.
 * The next log rotation is scheduled for the beginning of the next
   month. As it is still October, the next log rotation is configured
   for midnight, November 1st.
 * 80ms before midnight, we're done, and once again find spare time,
   so we decide to run the log rotation event ahead of schedule.
 * Once again, it is still October, so the next log rotation is November
   1st.
Systems with little to do has been observed executing this loop 18
times. Busier systems possibly have less spare time, but they only need
to rotate two times before all their log data is gone.

The effect on check bursts should not be noticeable - checks are
still scheduled by the second Nagios started, and the microsecond the
last check completed, and executed as close to this as possible. It's
just that the earliest check time is moved by 0.1 second to when the
event is scheduled.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>