test_checks: make this suite compile and enable it to run
might as well link to the stuff we need defined since we don't mock
most (any?) of it anyway. This way we don't have to do the whole
double maintenance dance.

Also took the opportunity to remove an unused variable and make the
Makefile build our dependencies when they're needed.

Signed-off-by: Anton Lofgren <alofgren@op5.com>