test_checks: disable test case for service state type transition
Rationale (also as inline comment):

Since we now (as per a466b8f75d19819ca329069b555ea8c3fb872717) use only
asynchronous checks, we cannot get the routing result necessary to
determine the real state of the host before we're done evaluating the
service check result. This means that we'll always run max_attempts
checks for a service before we set it in a hard state.

The solution proposed to mitigate this issue is to add some kind of
trigger to checks, such that (in this case) the scheduled host check,
when finished would update the state type of the service appropriately.

Change-Id: I3bd7208118c1ddb353793d3f3e2e950d61e2ec3b
Signed-off-by: Anton Lofgren <alofgren@op5.com>