Permanently disable the version check
Having an opensource daemon call in to a company server and report
which version it's running is just plain dumb. It's also probably
illegal in a lot of countries, as it could be considered spying.

Since it didn't really work anyway (3.99 was considered an earlier
version than 3.5.0), we might as well remove it.

This also mean we can remove of the networking code in netutils.[ch],
which means a net codeline reduction of 305 lines. Neato :)

While at it, we remove the editor-backed-up config.h.in~ file.

Signed-off-by: Andreas Ericsson <ae@op5.se>