Update deprecated attributes in sample config
* Changed normal_check_interval to check_interval
* Changed retry_check_interval to retry_interval

According to the documentation the old attribute names should
no longer exist, so switching to the newer documented behavior.

It should not change the behavior of a running environment, but it
can cause a lot of confusion when there is a mix of check_interval
and normal_check_interval in the same configuration object.