workers: Don't spam-reap children
Previously we used to try reaping children immediately after killing
them. When that fails (as it normally does when we try to reap half
a microsecond after sending the SIGKILL), the idea was to reschedule
a reaping 0.25 seconds later. However, we kinda failed that, since we
set the tv.tv_usec variable to 250000 instead of adding that value to
it.

This patch corrects it, inserting the missing plus-sign once again.

Signed-off-by: Andreas Ericsson <ae@op5.se>