Get rid of all the #ifndef NSCORE and #ifdef NSCGI code
Since we killed off the cgi's in d2cdcb490c1928, we can finally
also lay this horrible spaghetti of ifdefs to rest. The core is
now the core, and we don't have to keep track of which variables
of which structs are only ever available when we build a certain
set of binaries as opposed to another set.

Signed-off-by: Andreas Ericsson <ae@op5.se>