Allow check commands to be commandline escaped
If check commands are being used as arguments to notification
commands so that the user being notified can see exactly the
command that caused the state to change and that check command
has arguments containing single or double-quotes, there's no
sane way of using it as an argument without escaping it.

This patch allows such escaping for host and service check
commands, which should clear up the issue.

Signed-off-by: Andreas Ericsson <ae@op5.se>