Micro-optimize flapping detection
According to Nagios' documentation, flap_detection_options counts
which states we should care about when doing flapping detection.
Disregarding ignored states early means we can return earlier than
we would otherwise be able to (as states have to be tracked when
flapping is just disabled, but not when the object is in a state
that should be expressly ignored).

Signed-off-by: Andreas Ericsson <ae@op5.se>