Further flapping refactoring
If we shouldn't update state history, we can bail out early and
save a few cycles, since we know the flapping state won't change
since last time we updated it and therefore can't trigger any
action anyway.

If flapping detection is disabled, either globally or for the object
in question (but not explicitly for the object's current state; See
previous commit), we bail out immediately after calculating the
flapping state, since we still have to present the correct figures
in the user interface, but don't have to bother with notifications
and suchlike.

Signed-off-by: Andreas Ericsson <ae@op5.se>