lib/worker: Avoid logspam about unreapable children on 32-bit systems
On 32-bit systems, we shave off a whole lot of precision from the
scheduling queue microsecond timer, which means small increments
aren't necessarily timed properly. This in turn means that 32-bit
systems can still get log-spam of the following form:
  wproc <pid>: Failed to reap child <pid>: Some Error Message.

This patch should prevent that. As we now increment the timer by a
full second (the first time), the logspam at least won't be so
noticeable.

Signed-off-by: Andreas Ericsson <ae@op5.se>