worker: Get event runtime from the event runtime
Rather than trying to reconstruct it based on incomplete sources,
actually read the correct information at once.

"Incomplete" in this case means we never realized we'd tried to move the
event a few seconds forward, and just spammed the user over and over
with "failed to reap" messages.

This fixes bug op5#8112. Actually.

Reproducing said bug is a matter of using the following plugin:
  if [ $1 -lt 5 ]; then
          bash $0 $(($1 + 1))
  fi
  nc -l 8000
for any check, and simply wait 60 seconds for it to run.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>