objects: Fix object cache for host flap, stalk
The character for the UP state in both flap_detection_options and
stalking_options is 'o' - 'u' means unreachable. Nagios reads this
correctly, but saves the object.cache incorrectly. This makes CGIs
look wrong, but is even more serious in combination with merlin, which
uses the object.cache, not the config, to push config to pollers. This
makes flap detection and stalking completely broken for poller systems.

Hence this fixes bug op5#8049.

Sidenote: I'm getting sick of these "we generated config we couldn't
read" bugs.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>