commands: Disconnect command worker during restart
The command worker lives on when nagios restarts. This is quite nice -
your code can keep spewing in passive checkresults and whathaveyou,
even when nagios isn't actively monitoring.

"But what happens with those check results?", one might ask. One would,
in fact, be asking a very relevant question. Before this patch, they
would be submitted while nagios waits for workers to connect, but before
completely loading the configuration, leading to bug op5#8169 where nagios
would segfault. After this, the core process simply ignores the command
worker until the time comes when it would, during startup, be starting
the new command worker, after which it reads any incoming commands and
processes them ASAP. No more segfaults for you, fixing bug op5#8169.

This could be enhanced further, by letting the command worker buffer
incoming commands explicitly - instead, it simply lets the kernel buffer
them in it's tiny-but-for-a-few-seconds-hopefully-sufficient buffer.

As this is a race, it's hard to reproduce - I added a sleep(10) at the
bottom of the main loop in nagios.c while testing.

Change-Id: I5c0e6fe936c76ec44c7891190e8c5048c5319ee4
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>