commands: Let positive values of sticky mean yes
The acknowledge commands takes a "sticky" parameter. Before this patch,
0, 1, and >3 meant "not sticky", while 2 meant "sticky". I know why this
happened - internally, nagios uses 0 for "not acknowledged" and 1 for
"acknowledged", so internally, 2 is the first number available to use as
"stickily acknowledged". But. Really?

This changes the host and service acknowledge commands so that <=0 means
no and >0 means yes, the way all other booleans work. This does break
userspace API - though I suspect anyone using positive values other than
2 still intend for them to mean true, and are scratching their heads as
to why they don't work, as in op5#7794 or op5#5438.

Change-Id: Ib440df797b63c18d440cc1802bcce03b32341968
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>