initialize event queue earlier
neb modules (ex. pnp4nagios) register user events during initialization
(triggered by neb_load_all_modules()). In order to make that work, we need to
make sure the event queue is already initialized at that point. Unfortunately
initializing the event queue requires the number of objects to determince the
size of the queue, so reading the objects had to be rearranged too.

Signed-off-by: Sven Nierlein <sven@nierlein.de>