config parsing: Avoid using add_object_to_objectlist() where possible
In quite a lot of places, we don't care about ordering, so using
insert-last in our singly linked list without a tail pointer is
severely suboptimal.

This patch reduces the O(n*(n+1)/2), which would cost 5000050000
list operations for each object that attaches to 100000 other
objects, to O(n), which costs 100000 list operations for the same
effect.

Signed-off-by: Andreas Ericsson <ae@op5.se>