lib/nsock: Make nsock_printf{_nul}() mimic printf() properly
They were always intended to, but since they were backed by a function
using a static size buffer, we could never print more than 4096 bytes
in a single printf(). This patch amends that, and although it will
be slightly slower to use the generic function, we now know we'll
never run into buffer overflows when adding the terminating nul-byte.
Failures will only ever happen when we fail to allocate memory, in
which case the call is scrapped in its entirety.

In exceptional circumstances, this could otherwise lead to information
leaks if modules use the nsock_printf() functions to send information
to listeners that exceeded 4096 bytes per write.

Discovered-by: Aman Gupta <aman@tmm1.net>
Reported-by: Dirkjan Bussink <dirkjan.bussink@github.com>
Thanks-to: Markus Frosch <markus@lazyfrosch.de>
Signed-off-by: Andreas Ericsson <ae@op5.se>