Replace the COPYING file
Nagios, a GPLv2 (or later?) project, had its license in a LICENSE file.
The autotoolsification auto-inserted a GPLv3 license into the GNU
standard COPYING file, which means depending on what you're used to, the
license seems to vary.

Drop the LICENSE file and use the expected text in the COPYING file
instead.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>