checks.c: cleaned `reschedule failed scheduled service checks'
if a given scheduled service check failed by any reason we
need to reschedule it only if it is supposed to be scheduled.

the old code works like a charm, despite the fact it looks
a little bit bizarre. imho new code gets much more cleaner this way,
no functional changes.

Signed-off-by: Ricardo Maraschini <ricardo.maraschini@gmail.com>