Handle retention data gracefully when config changes
Previously, object settings changed by submitting external commands
were impossible to change via config, as the modified_attributes
variable would always trigger, and always be read in to see which
variables we should retain.

This has led to much confusion from people who have, for example,
disabled active checks through external commands, then changed
the settings for those active checks through configuration only to
notice that no active checks are run.

With this patch, we stash the configured object the first time we
change it and only print the changed variables to the retention
file, alongside the then-configured value. This means that we now
normally write less data in the retention data file, but when there
are things we actually need to remember we'll also have the config
for that setting at the time when the external command was submitted
to change the value. When we then read retention data back in, we can
compare the configured value with the last configured value and see
if the config has changed during the time Naemon was shut down. If it
has, we will use the new configuration value rather than the retained
settings. If it hasn't, we'll use the retained data.

So far, this patch only covers boolean variables for hosts and
services. Since we've disabled changing check- and eventhandler
commands, that covers most of it though (and only timeperiod settings
remain to be handled).

Contacts should be fixed in the same way, and will be soon.

Signed-off-by: Andreas Ericsson <ae@op5.se>