retention data: Prevent coredumps and memory leaks
Previously when reading modified timeperiods from objects, we would
free() the name of the timeperiod from the object and strdup() in a
new one. However, since quite some time now, we don't even copy the
name to the object anymore, but just assign it from the name used
for the timeperiod itself.

This patch brings the retention data code in line with the config
parsing routines, preventing coredumps (although only on debug
messages or when the same timeperiod has been changed for multiple
objects through external commands without a program restart in
between) and memory leaks (on soft restarts when a timeperiod has
changed for a single object).

Signed-off-by: Andreas Ericsson <ae@op5.se>