checks: Be consistent in handling missing output
For service checks, we let the plugin_output be "(No output returned
from the plugin)" in case we get _no_(NULL) short output. For host
checks, this is also done when the output is empty("")(albeit with the
string "(No output returned from host check)" instead).

This patch changes the behaviour for host checks, so that we only
replace the output if it is NULL, since this signifies a fault.

If the output is an empty string, this is likely the intended - though,
one could argue, misdirected - behaviour by the plugin author, and I
don't think it's appropriate that we mess around with it.

Signed-off-by: Anton Lofgren <alofgren@op5.com>