naemon/utils.c: removed useless code on my_system_r()
when parent receives child return it just ignores if
exit status is > than 3 or < -1, so it makes no sense
to sum 128 to the received signal number generated by
the plugin.

btw, in the way the code was it always ignore signals
on plugins anyway due to overwrite of `result' value
on the next line.

we verify for timeout a few lines below also, but it
is not the plugin execution who timed out, but the
child process, with SIGALRM, so again: the plugin is
not signaled by us there, we just call kill() a few
lines below in case it stucked.

Signed-off-by: Ricardo Maraschini <ricardo.maraschini@gmail.com>