defaults.h: Specify autotools macros compile-time
Because autotools is autotools, the default paths while configure is
being executed can contain macro variables such as "${prefix}". This
variable is resolved by make, meaning that it will be available in a
final-ish state compile-time (even though it *could* be changed
install-time, but at that time we're not allowed to recompile anything
and by the way what exactly do we think we're doing compiling paths into
the binary we should be ashamed of ourselves we violate gnu guidelines
all over - but, you know, whatever) so that's where we provide it to
defaults.h.

Signed-off-by: Robin Sonefors <ozamosi@flukkost.nu>