test_macros: Escape {host,service}group note macros
This is pretty much identical to the patch in 83d1046dfb6 (which
fixes op5#8266) and as that commit implies, this is a nasty way of dealing
with the issue. There seems to be something fundamentally wrong with
the way macros work today, and we should probably spend some time
working out a better solution.

Anyway, I also added a test which should cover regressions of the service
part of op5#8266, which is the same issue as op5#8270, but for service and host
notes.

This fixes op5#8270.

Change-Id: Ic382c07b1aee988b01234bccdcd09e2d8d99a910
Signed-off-by: Anton Lofgren <alofgren@op5.com>