tests: Fix undefined reference to `ck_assert_ptr_eq`
The ck_assert_ptr family of functions were added in check 0.9.10, which
unfortunately is newer than the versions available in the current stable
releases of Debian and Ubuntu. We workaround this by replacing calls to
`ck_assert_ptr` with `ck_assert`. The failure messages won't be as
pretty but at least the tests run...

Signed-off-by: Adam James <adam.james@transitiv.co.uk>