Wrap exported code in NAGIOS_{BEGIN,END}_DECL
The macros in question does nothing more advanced than conditionally
wrapping the code in an extern "C" {} block. Those are really important,
though, as C++ code (such as livestatus) will be unable to use code
exported without these wrappers.

A bit ironic (in the Alanis Morissette sense), really, that the lib code
is the only code that was missing these.

This makes it possible to use nsock in livestatus code, making it
easy to replace the current racy external command code in livestatus
with a wrapper around the qh "command" handler, making it easy for Thruk
to submit passive check results without crashing mod_gearman/merlin. And
thus the Yak has been shaved!

Signed-off-by: Robin Sonefors <ozamosi@flukkost.nu>