naemon/utils: Don't get matching time range for the wrong day
So, setting SECS_PER_DAY to the correct value caused the time range
matching to get the time range for the wrong day (+1).

On further inspection, we shouldn't really add a (diurnal) day here,
at all. So, I'm removing it. Previously, it'd work because the value
would be rounded down so that we'd still get the correct day anyway.

Found another use for the SECS_PER_DAY macro, though, so there's that.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>