t-tap: Fix make distcheck again
There were issues involving the whole "you can't write into the srcdir,
and it's annoying to put random files into builddir", and there were a
couple of fuckups on my part with regards to which one of the two to
use.

Now it passes on my laptop, at least - travis, you're up!

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>