Don't initialize dynamic globals with static memory
reset_variables initializes a bunch of global variables. Among them
are many strings, which are strdup()ed from the static constant that
contains their default value. When parsing a config file, the ones with
an explicit value are free()d and replaced with an strdup()ed version of
what was in the config file.

If reset_variables haven't yet been called, some of the variables would
be initialized to the static constant that is their default value, so
any attempt to load a config file would lead to a segfalt, as the static
constant was attempted to be free()d. Instead, consistently initialize
the variables to NULL until reset_variables has been called.

While code *is* supposed to call reset_variables, this makes things a
bit less unpredictable if it doesn't.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>