config parsing: Consider services with 'name' potential templates
Weird commit subject, but whatever. With this patch, we no longer
bomb out when encountering services that lack both 'host_name' and
'hostgroup_name' variable so long as they have a 'name' variable
and can thus be used as templates.

This brings config parsing in line with how Nagios does it, which
is a good thing for a project supposed to be a drop-in replacement.

Signed-off-by: Andreas Ericsson <ae@op5.se>