Avoid calling update_program_status() from logging.c
It's only ever done when rotating the logfile, and the call to do so
is only done from events.c, where we now call update_program_status()
after the logfile has been rotated.

Avoiding it in logging.c means one can link to logging.o without
having to pull in statusdata.o and, by extension, xsddefault.o.
This has no impact what so ever on functionality, but makes it a lot
more pleasant to build tests and to create tools that utilize the
naemon core code for other uses.

Signed-off-by: Andreas Ericsson <ae@op5.se>