base/workers: Consistensify
When the check output is empty, we try to print the stderr, even if the
stderr is empty, which results in confusing output - "(No output on
stdout) stderr:" (remember: nagios strips/chomps output) - which gives
the reader the impression that there's a stderr to read somewhere.

Instead, when there's not anything on stderr, set the output to NULL,
which is passed to process_check_result which is passed to
handle_async_host_check_result which is passed to parse_check_output
which results in check_output being set to NULL which
handle_async_host_check_result then replaces with "(No output returned
from host check)". Because layering.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>