Make it a bit easier to build addon projects against naemon headers
The previous defaults.h file included autotool macros that were only
set as preprocessor macros on the command line, which made defaults.h
unusable from addon projects.

With this patch we add a buildopts.h, where projects that don't want
to link against the naemon sources but still want to know which
defaults to use (for whatever reason) can #include the defaults.h
file and get the compile-time defaults.

Signed-off-by: Andreas Ericsson <ae@op5.se>