Fix a bunch of compiler warnings and annoyances with tests
The indentation was pretty horrible, and the plethora of warnings
was just appalling. Naming variables the same as types is not only
frowned upon, but an actual error for many compilers, and mixing
declarations and statements is illegal in ISO C90.

The rest is mostly just getting rid of shadowed variables and
functions and getting included files in the right order.

Signed-off-by: Andreas Ericsson <ae@op5.se>