t-tap/smallconfig/retention.dat: Make retained host downtime sensible
It doesn't make sense to have zero-second downtimes, and the code we
have today rightly denies such crap as nonsense. However, the tests
in test_config.c (well, soon-to-be test_config.c) attempts to locate
a host downtime with exactly those parameters, stashed in our
retention data testfile.

This patch fixes it by nudging the end_time by 10 seconds, thus
allowing the downtime a small piece of space in the fourth dimension
in where it can, if not thrive, then at least exist.

Signed-off-by: Andreas Ericsson <ae@op5.se>